fix: move private test to inside module
